Continue updating documentation
