+++ docs/autodoc/label_metadata.rst
@@ -6,3 +6,7 @@ These descriptions are used to create the practical python classes in ``schemas.
 
 See the source code for the dicts themselves. 
 See the :ref:`Training from Scratch <training_from_scratch>` guide for an explanation of what these are and how they related to ``schemas.py`` and the multi-question loss function.
+
+|
+
+.. autofunction:: zoobot.shared.label_metadata.extract_questions_and_label_cols
