+++ docs/guides/representations.rst
@@ -23,10 +23,10 @@ See :ref:`training_from_scratch` for a guide to training a new model.
 
 We have published pretrained weights for models trained on GZ DECaLS - see :ref:`datanotes`. 
 You could start with these and calculate representations on some new galaxies.
-See ``make_predictions_loop.py`` for how to load the weights, and below for how to calculate representations.
+See `zoobot/tensorflow/examples/make_predictions_loop.py <https://github.com/mwalmsley/zoobot/blob/pytorch/zoobot/tensorflow/examples/make_predictions_loop.py>`__ for how to load the weights, and below for how to calculate representations.
 
 You might also want to start from a pretrained model and use finetuning to get the best representation for your problem.
-See ``finetune_advanced.py`` for an example. This adds some complexity, so we suggest trying with our pretrained weights first.
+See `zoobot/tensorflow/examples/finetune_advanced.py <https://github.com/mwalmsley/zoobot/blob/pytorch/zoobot/tensorflow/examples/finetune_advanced.py>`__ for an example. This adds some complexity, so we suggest trying with our pretrained weights first.
 
 Extracting the Representation
 -----------------------------
@@ -45,10 +45,20 @@ This configures the model like so:
 
 As always, remember to check ``run_name`` and any file paths.
 
-``make_predictions_loop.py`` will then save the representations for each galaxy to files like {run_name}_{index}.csv.
-These files are a bit awkward as they include lots of numbers like ``[[0.4, ...]]``.
-Remove the brackets with ``predictions/reformat_predictions.py``.
+``make_predictions_loop.py`` will then save the representations for each galaxy to files like {run_name}_{index}.hdf5.
+hdf5 files are simply numpy arrays saved to disk, sometimes with associated metadata (here, the ``label_cols`` used) - see `here <https://docs.h5py.org/en/stable/quick.html>`__  for more details.
 
-Finally, compress the 1280-dim representation into a lower dimensionality using PCA with ``representations/compress_representations.py``.
+Compressing the Representation (optional)
+-----------------------------------------
+
+The representation will have 1280 features per galaxy (per forward pass - see note below).
+This representation is typically highly correlated and can be well-described using fewer features.
+I strongly suggest compressing the 1280-dim representation into a lower dimensionality using PCA (or similar).
+I suggest `Incremental PCA <https://scikit-learn.org/stable/auto_examples/decomposition/plot_incremental_pca.html>`__ for datasets larger than about 50,000 rows.
 The compressed representation is mathematically very similar (PCA should preserve most of the interesting variation) but much easier to work with.
+`zoobot/shared/compress_representations.py <https://github.com/mwalmsley/zoobot/blob/pytorch/zoobot/shared/compress_representations.py>`__ shows an example.
 
+.. note:: 
+
+    If you are making multiple forward passes for galaxy, your saved representation will be 3-dimensional: (galaxy, feature, forward pass).
+    You will need to choose how to reduce this to 2 dimensions so that you can apply PCA - perhaps by taking a mean, or simply picking the first pass.
