--- zoobot/tensorflow/representations/README.md
@@ -1,3 +1,0 @@
-# Representations Guide
-
-Please see [https://zoobot.readthedocs.io/usage/representations.html]
